+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changelog
 
 * Using secure links for documentation, thanks @[Phylon](https://github.com/Phylon)
 * Made reboot messages translatable ([issue](/../../issues/1580))
+* Limited usage data to 12 hours or 700 lines
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)